fix(subagents): satisfy rebased CI checks
